included some changes for PR
