+++ seisbench/models/__init__.py
@@ -1,7 +1,7 @@
 from .aepicker import BasicPhaseAE
 from .base import GroupingHelper, SeisBenchModel, WaveformModel, WaveformPipeline
 from .cred import CRED
-from .deepdenoiser import DeepDenoiser, SeisDAE
+from .deepdenoiser import DeepDenoiser
 from .depthphase import DepthFinder, DepthPhaseModel, DepthPhaseNet, DepthPhaseTEAM
 from .dpp import DeepPhasePick, DPPDetector, DPPPicker
 from .eqtransformer import EQTransformer
@@ -10,5 +10,6 @@ from .lfe_detect import LFEDetect
 from .obstransformer import OBSTransformer
 from .phasenet import PhaseNet, PhaseNetLight, VariableLengthPhaseNet
 from .pickblue import PickBlue
+from .seisdae import SeisDAE
 from .skynet import Skynet
 from .team import PhaseTEAM
